v3.2: Binance WS, sprite-based rendering, disable double-buffer
Major changes:
- Add Binance WebSocket for real-time BTC price + 1m kline backfill
- All UI draws through persistent zone sprites (Z0/Z1/Z2) — no direct tft
- Toast and flash feedback rendered inside sprites (eliminates tearing)
- Dirty zone tracking for selective buffer sync
- Disable double buffering (was causing vertical bounce/repeat artifacts)
- Add OHLC candlestick chart, chart markers overlay, dollar carousel
- Add sparkline/dollar morph animations (disabled on RGB panel)
- Price flash animation (green/red fade on WS price change)
- Partial price-only Z1 update for efficient counter animation
- Swipe gestures for period/chart style navigation

+++ src/animation.h
@@ -70,21 +70,21 @@ public:
 class SparklineAnimator {
     unsigned long startMs_;
     uint16_t durationMs_;
-    uint8_t totalCount_;
+    uint16_t totalCount_;
     bool animating_;
 
 public:
     SparklineAnimator() : startMs_(0), durationMs_(800),
                           totalCount_(0), animating_(false) {}
 
-    void start(uint8_t count, uint16_t ms = 800) {
+    void start(uint16_t count, uint16_t ms = 800) {
         totalCount_ = count;
         startMs_ = millis();
         durationMs_ = ms;
         animating_ = true;
     }
 
-    uint8_t visibleCount() {
+    uint16_t visibleCount() {
         if (!animating_) return totalCount_;
 
         unsigned long elapsed = millis() - startMs_;
@@ -96,7 +96,7 @@ public:
         float t = (float)elapsed / durationMs_;
         // Ease-out quad for smooth reveal
         float ease = 1.0f - (1.0f - t) * (1.0f - t);
-        uint8_t visible = (uint8_t)(ease * totalCount_);
+        uint16_t visible = (uint16_t)(ease * totalCount_);
         return visible < 1 ? 1 : visible;
     }
